Finalized the name space and the easy_install egg changes
The Makefile.in is now adapted to the changes that the name space
introduction caused (new subdirectory). The easy_install egg
setup file needed to know about the version number, so
it was time to ditch the py_interface.pub file (which previously
held the version number), and move the version number to a vsn file.
and Also, in the process, I (re)created a ChangeLog file, to
completely cover everything (and more) that was in the .pub file.

(The py_interface.pub file was needed to get things published
on the Erlang User contribution page http://www.erlang.org/user.html,
but that's outdated by now.)