Handle eLen in PORT2_RESP that isn't 2 bytes
As of erlang 13B04, I started seeing an eLen of 0, which caused
the ReadInt2 to traceback.

This is the minimal change needed to keep the code working while
not breaking the implicit assumption that a 16-bit integer may
be coming in the extra field.

See commit a841673 in the erlang/otp tree:
http://github.com/erlang/otp/commit/a841673669604f5be838a0a0042f05e18ae3827c