Add __eq__, __ne__ and __hash__ to Erl<Type> classes
The __eq__ and __ne__ are so the instances can be compared, as a
precursory step to them being keys in a map type. Also, since we
add __eq__ and __ne__, we need to add __hash__ if we want them to be
hashable. At least pids need to be hashable by the erl_node.

While at it, fix comparison to check the 'other' element is an
instance of the right class, before checking the 'other's members.